Add advanced option to prevent loading kernel modules
Add a --skip-module-load option to the installer, which will
directly set the option variable of the same name. The new option
complements the existing --skip-module-unload option.

In previous versions, the installer would always try to load the
kernel module if it was built for the running kernel. This
operation fails when no NVIDIA hardware is present, preventing
installation.

The new option suppresses all module loading, everywhere. It
allows installation to complete on systems which do not have
NVIDIA hardware connected at install-time.